// src/config/enums.ts
export enum ErrorType {
  QUERY_ERROR = "QUERY_ERROR",
  NETWORK_ERROR = "NETWORK_ERROR",
  VALIDATION_ERROR = "VALIDATION_ERROR",
  GEOMETRY_ERROR = "GEOMETRY_ERROR",
}

export enum PropertyActionType {
  SET_ERROR = "PROPERTY_WIDGET/SET_ERROR",
  CLEAR_ERROR = "PROPERTY_WIDGET/CLEAR_ERROR",
  SET_SELECTED_PROPERTIES = "PROPERTY_WIDGET/SET_SELECTED_PROPERTIES",
  CLEAR_ALL = "PROPERTY_WIDGET/CLEAR_ALL",
  SET_QUERY_IN_FLIGHT = "PROPERTY_WIDGET/SET_QUERY_IN_FLIGHT",
  SET_RAW_RESULTS = "PROPERTY_WIDGET/SET_RAW_RESULTS",
  REMOVE_WIDGET_STATE = "PROPERTY_WIDGET/REMOVE_WIDGET_STATE",
}